refactor: move function `sign_in` into `setting`
