+++ src/rust/zh.boylove/src/lib.rs
@@ -7,20 +7,14 @@ use aidoku::{
 	error::{AidokuError, AidokuErrorKind, Result},
 	helpers::substring::Substring,
 	prelude::*,
-	std::{
-		defaults::defaults_get, html::unescape_html_entities, json, net::Request, String, ValueRef,
-		Vec,
-	},
+	std::{html::unescape_html_entities, json, net::Request, String, ValueRef, Vec},
 	Chapter, DeepLink, Filter, Listing, Manga, MangaContentRating, MangaPageResult, MangaStatus,
 	Page,
 };
 use alloc::{borrow::ToOwned as _, string::ToString};
-use base64::{engine::general_purpose, Engine};
 use helper::{
-	setting::change_charset,
-	url::{
-		ChapterQuery, DefaultRequest as _, Index, LastUpdatedQuery, Url, CHAPTER_PATH, MANGA_PATH,
-	},
+	setting::{change_charset, sign_in},
+	url::{ChapterQuery, DefaultRequest as _, Index, LastUpdatedQuery, Url},
 	MangaList as _, MangaListRes as _, Part, Regex,
 };
 
@@ -328,38 +322,6 @@ fn get_content_rating(categories: &[String]) -> MangaContentRating {
 		return MangaContentRating::Safe;
 	}
 	MangaContentRating::Nsfw
-}
-
-fn sign_in() -> Result<()> {
-	let captcha = defaults_get("captcha")?.as_string()?.read();
-
-	let is_wrong_captcha_format = captcha.parse::<u16>().is_err() || captcha.chars().count() != 4;
-	if is_wrong_captcha_format {
-		let sign_in_page = Url::SignInPage.get().html()?;
-
-		let captcha_img_path = sign_in_page.select("img#verifyImg").attr("src").read();
-		let captcha_img = Url::Abs {
-			path: &captcha_img_path,
-		}
-		.get()
-		.data();
-		let base64_img = general_purpose::STANDARD_NO_PAD.encode(captcha_img);
-
-		return Ok(println!("{}", base64_img));
-	}
-
-	let username = defaults_get("username")?.as_string()?.read();
-	let password = defaults_get("password")?.as_string()?.read();
-	let sign_in_data = format!(
-		"username={}&password={}&vfycode={}&type=login",
-		username, password, captcha
-	);
-
-	let response_json = Url::SignIn.post(sign_in_data).json()?;
-	let reponse_obj = response_json.as_object()?;
-	let info = reponse_obj.get("info").as_string()?;
-
-	Ok(println!("{}", info))
 }
 
 trait Parser {
